D 26/04/2021 Done

+++ app/Http/Controllers/MapsController.php
@@ -4,14 +4,18 @@ namespace App\Http\Controllers;
 
 use App\Models\Maps;
 use Illuminate\Http\Request;
+use Bitfumes\Multiauth\Model\Admin;
+use Auth;
 
 class MapsController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    // public function __construct()
+    // {
+        // $this->middleware('auth:admin');
+        // $this->middleware('role:user')->only('index');
+    // }
+
+    
     public function index()
     {
         $maps = Maps::get();
@@ -36,6 +40,11 @@ class MapsController extends Controller
      */
     public function store(Request $request)
     {
+        if(Auth::user()->roles->first()->name != 'user'){
+            return redirect('/index')->with('status', 'Access Denied!');
+
+        }
+        
         $data = [
              'type' => $request->type,
              'proponent' => $request->proponent,
@@ -64,6 +73,7 @@ class MapsController extends Controller
      */
     public function show(Maps $maps)
     {
+        // dd($maps);
        return view('dashboard.marker')->with('maps',$maps);
     }
 
@@ -75,6 +85,11 @@ class MapsController extends Controller
      */
     public function edit(Maps $maps)
     {
+        if(Auth::user()->roles->first()->name != 'user'){
+            return redirect('/index')->with('status', 'Access Denied!');
+
+        }
+        
         return view('dashboard.edit')->with('map',$maps);
     }
 
@@ -87,6 +102,11 @@ class MapsController extends Controller
      */
     public function update(Request $request, Maps $maps)
     {
+        if(Auth::user()->roles->first()->name != 'user'){
+            return redirect('/index')->with('status', 'Access Denied!');
+
+        }
+
         $input = $request->all();
         // dd($input);
         $maps->update($input);
@@ -102,6 +122,11 @@ class MapsController extends Controller
      */
     public function destroy(Maps $maps)
     {
+        if(Auth::user()->roles->first()->name != 'user'){
+            return redirect('/index')->with('status', 'Access Denied!');
+
+        }
+
         $maps->delete();
         $status = 'Map Deleted';
         return back()->with(['status' => $status]);
